Remove `DispatchToSpecializedApply` phase
